requirements.txt
annotated-types @ file:///C:/miniconda3/conda-bld/annotated-types_1761745023123/work
anyio @ file:///C:/miniconda3/conda-bld/anyio_1758622432793/work
asttokens==3.0.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
coloredlogs @ file:///home/conda/feedstock_root/build_artifacts/coloredlogs_1733927977068/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
cycler==0.12.1
dateparser==1.2.2
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1754523437/work
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi @ file:///C:/miniconda3/conda-bld/fastapi_1757101273602/work
filelock==3.19.1
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1739279199749/work
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
humanfriendly @ file:///D:/bld/humanfriendly_1733927963288/work
humanize==4.13.0
idna @ file:///C:/miniconda3/conda-bld/idna_1761912043328/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipdb==0.13.13
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.4.2
matplotlib-inline==0.2.1
maya==0.6.1
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.2.1
ninja==1.13.0
numpy==1.22.4
onnx @ file:///D:/bld/bld/rattler-build_onnx_1751877496/work
onnxruntime @ file:///C:/bld/onnxruntime_1730051925342/work/build-ci/Release/dist/onnxruntime-1.19.2-cp39-cp39-win_amd64.whl#sha256=bcf85962add5d45215f7347532fd717ce4c4e7ecf8ed16f9bcc130613417c108
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
outcome==1.3.0.post0
packaging==25.0
pandas==1.5.1
parso==0.8.5
pendulum==3.1.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
playwright==1.55.0
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1740663155517/work
pure_eval==0.2.3
pycparser==2.23
pydantic @ file:///C:/miniconda3/conda-bld/pydantic_1760697582034/work
pydantic_core @ file:///C:/miniconda3/conda-bld/pydantic-core_1760597458296/work
pyee==13.0.0
Pygments==2.19.2
pymel==1.4.0
pyparsing==3.2.5
pyreadline3 @ file:///D:/bld/pyreadline3_1749148201214/work
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1752564076/work
PyYAML @ file:///C:/b/abs_e9vlazjouh/croot/pyyaml_1681203360243/work
pyzmq @ file:///D:/bld/pyzmq_1755799563460/work
regex==2025.11.3
requests==2.32.5
scipy==1.6.2
selenium==4.36.0
six==1.17.0
snaptime==0.2.4
sniffio @ file:///C:/b/abs_3akdewudo_/croot/sniffio_1705431337396/work
sortedcontainers==2.4.0
stack-data==0.6.3
starlette @ file:///C:/miniconda3/conda-bld/starlette-recipe_1756910711216/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946057943/work
tensorboardX==2.6
tomli==2.3.0
tornado @ file:///D:/bld/tornado_1754732005081/work
tqdm==4.61.2
traitlets==5.14.3
trimesh==3.12.7
trio==0.31.0
trio-websocket==0.12.2
typing-inspection @ file:///C:/miniconda3/conda-bld/typing-inspection_1760614191697/work
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
